03_everything_is_an_api/07_fastapi/04_fastapi_project_with_dependencies/requirements.txt
fastapi
uvicorn
pydantic
passlib[bcrypt]         # For password hashing
python-jose[cryptography] # For JWT (JSON Web Tokens)
openai                  # For interacting with OpenAI's API
asyncpg                 # For asynchronous PostgreSQL database interaction (recommended for FastAPI)
python-multipart